PMM-6091 Refactoring of usePerformUpdate
